simplified graph crawling in lieu of new links.json
